이름 중복 로그인 여부 메세지 추가.

+++ src/routes/Character.routes.js
@@ -14,6 +14,14 @@ router.post("/character-up", Middleware, async (req, res) => {
   const userId = req.user.userId;
 
   try {
+    const existingCharacter = await prisma.character.findUnique({
+      where: { name },
+    });
+
+    if (existingCharacter) {
+      return res.status(400).json({ message: "이미 존재하는 닉네임 입니다." });
+    }
+
     const character = await prisma.character.create({
       data: {
         name,
@@ -29,11 +37,28 @@ router.post("/character-up", Middleware, async (req, res) => {
 // 캐릭터 삭제
 router.delete("/character/:id", Middleware, async (req, res) => {
   const { id } = req.params;
+  const userId = req.user ? req.user.userId : null;
 
   try {
+    if (!req.user) {
+      return res.status(401).json({ message: "로그인이 필요합니다." });
+    }
+
+    const character = await prisma.character.findUnique({
+      where: { characterId: +id },
+    });
+
+    if (!character) {
+      return res.status(404).json({ message: "존재하지 않는 캐릭터입니다." });
+    }
+
+    if (character.userId !== userId) {
+      return res.status(403).json({ message: "본인의 캐릭터가 아닙니다." });
+    }
     await prisma.character.delete({
       where: { characterId: +id },
     });
+
     res.status(200).json({ message: "캐릭터가 정상적으로 삭제되었습니다." });
   } catch (error) {
     res.status(400).json({ message: "캐릭터 삭제에 실패하였습니다." });
@@ -65,6 +90,7 @@ router.get("/character/:id", async (req, res) => {
 // 로그인했을 때 캐릭터 조회
 router.get("/my-character/:id", Middleware, async (req, res) => {
   const { id } = req.params;
+  const userId = req.user ? req.user.userId : null;
 
   try {
     if (!req.user) {
@@ -73,20 +99,14 @@ router.get("/my-character/:id", Middleware, async (req, res) => {
 
     const character = await prisma.character.findUnique({
       where: { characterId: +id },
-      select: {
-        characterId: true,
-        name: true,
-        gold: true,
-        attack: true,
-        str: true,
-        dex: true,
-        int: true,
-        luk: true,
-      },
     });
 
     if (!character) {
       return res.status(404).json({ message: "캐릭터를 찾을 수 없습니다." });
+    }
+
+    if (character.userId !== userId) {
+      return res.status(403).json({ message: "본인의 캐릭터가 아닙니다." });
     }
 
     res.status(200).json({ character });
